+++ app/modules/database/index.php
@@ -51,7 +51,7 @@ $config = [
 
         Type::overrideType(Types::SIMPLE_ARRAY, "\GreenCheap\Database\Types\SimpleArrayType");
         Type::overrideType(Types::JSON, "\GreenCheap\Database\Types\JsonType");
-        Type::overrideType(Types::JSON_ARRAY, "\GreenCheap\Database\Types\JsonArrayType");
+        #Type::overrideType(Types::JSON_ARRAY, "\GreenCheap\Database\Types\JsonArrayType");
     },
 
     "autoload" => [